# Client/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Quick Start",
        "description": "<p>Documentation API of <b>Text Mining Project</b> has been written by Live Laugh Love.</p><p><b><i>This is an internal resource. Do not share it in any way.</i></b></p><p>You should read carefully the descriptions, response status codes and arguments request belong to each API Resource. It take your time, but you will avoid a lot of bugs afterward.</p><b>The application case sensitive.</b>",
        "contact": {
            "name": "nhudangstudy, anhhoangdev, cocaa",
            "email": "nhucnh21416c@st.uel.edu.vn"
        },
        "version": "3.0.45.156"
    },
    "servers": [
        {
            "url": "https://localhost:7065"
        }
    ],
    "paths": {
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryModel"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryModel"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "parameters": [
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCategoryModel"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCategoryModel"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCategoryModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/api/reviews/product/{asin}": {
            "get": {
                "tags": [
                    "Review"
                ],
                "parameters": [
                    {
                        "name": "asin",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews/user/{email}": {
            "get": {
                "tags": [
                    "Review"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews/featured": {
            "get": {
                "tags": [
                    "Review"
                ],
                "parameters": [
                    {
                        "name": "count",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 4
                        }
                    },
                    {
                        "name": "minimumRating",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "double",
                            "default": 4
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseModel"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reviews/{asin}": {
            "post": {
                "tags": [
                    "Review"
                ],
                "parameters": [
                    {
                        "name": "asin",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReviewRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReviewRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/user/{email}": {
            "post": {
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateUserRequestModel"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateUserRequestModel"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateUserRequestModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetByIdPublicAccountModel"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetByIdPublicAccountModel"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetByIdPublicAccountModel"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateCategoryModel": {
                "type": "object",
                "properties": {
                    "categoryName": {
                        "type": "string",
                        "nullable": true
                    },
                    "categoryDescription": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "CreateReviewRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "imageUrl": {
                        "type": "string",
                        "nullable": true
                    },
                    "starRating": {
                        "type": "number",
                        "format": "double"
                    },
                    "reviewContent": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "CreateUpdateUserRequestModel": {
                "required": [
                    "dateOfBirth",
                    "firstName",
                    "lastName"
                ],
                "type": "object",
                "properties": {
                    "firstName": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "lastName": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "additionalProperties": false
            },
...
            "GetByIdPublicAccountModel": {
                "required": [
                    "email",
                    "fullName"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "minLength": 1,
                        "type": "string",
                        "format": "email"
                    },
                    "fullName": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "ReviewResponseModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "asin": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "imageUrl": {
                        "type": "string",
                        "nullable": true
                    },
                    "starRating": {
                        "type": "number",
                        "format": "double"
                    },
                    "reviewContent": {
                        "type": "string",
                        "nullable": true
                    },
                    "ownedBy": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "UpdateCategoryModel": {
                "type": "object",
                "properties": {
                    "categoryName": {
                        "type": "string",
                        "nullable": true
                    },
                    "categoryDescription": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "The token to execute the private resources.",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}